+++ src/index.html
@@ -5,6 +5,8 @@
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
     <title>TodoList</title>
     <link rel="stylesheet" href="https://use.fontawesome.com/releases/v5.13.0/css/all.css"/>
+    <script type="module" src="index.js"></script>
+    <link rel="stylesheet" href="style.css">
 </head>
 <body>
