+++ src/modules/Tasks/Tasks.js
@@ -1,0 +1,16 @@
+
+import Status from '../TaskStatus/TaskStatus.js';
+import { addTodo } from '../addTodo/addTodo.js';
+import { removeTodo } from '../removeTodo/removeTodo.js';
+
+
+
+
+
+const Tasks = () => {
+  addTodo();
+  removeTodo();
+
+};
+
+export default Tasks;
