+++ src/modules/clear/clearAll.js
@@ -1,0 +1,6 @@
+const clearBtn = document.querySelector('.btn');
+
+      clearBtn.addEventListener('click', (e) => {
+        e.preventDefault();
+        renderTodo();
+      });
