+++ src/modules/consts/consts.js
@@ -1,0 +1,6 @@
+export const todoArray = [];
+export const sendToLocalStorage = () => {
+  localStorage.setItem('todoArray', JSON.stringify(todoArray));
+};
+
+
